refactor(desktop): split github.rs into a github/ module (#233)
github.rs was the largest Rust file (1107 LOC, 17 commands, 17 structs)
and the only complex backend module with zero tests. Split it into a
cohesive directory module with no behavior change and no change to the
public IPC surface — github/mod.rs re-exports every command, so lib.rs is
untouched and all github::* invoke paths still resolve.

- github/mod.rs — shared gh-client layer (hidden_command, GhErrorKind,
  classify_gh_error, gh_error_message, run_gh_with_retry) + submodule
  declarations and re-exports
- github/rate_limit.rs, issues.rs, pulls.rs, releases.rs, labels.rs,
  git.rs, auth.rs — domain split; ClaudeCli detection lives in auth.rs
- 13 new serde parse tests over gh --json fixtures (issues, pulls,
  releases, rate-limit, labels) — the module had zero tests before

Pure move: every #[serde(...)], #[tauri::command], doc comment and body
preserved verbatim; imports trimmed per file (clean under -D warnings).

cargo clippy --all-targets --deny warnings clean; cargo test green
(lib 42, including 13 new github tests).

Closes #233
